+++ CHANGELOG.md
@@ -7,6 +7,9 @@ Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
 
+### Added
+- Added --subject-template argument for setting the SNS message subject
+
 ## [0.4.0] - 2021-02-24
 
 ### Added